yc-workbank-pipeline/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# LLM APIs
anthropic>=0.25.0
openai>=1.30.0

# Google Sheets integration (optional)
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Web interface
flask>=2.3.0

# Data processing
lxml>=4.9.0
openpyxl>=3.1.0

# Utility
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0

# Development (optional)
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0
